requirements.txt
wheel
discord.py
uvloop
wolframalpha
osuapi
pyowm
ruamel.yaml
jishaku
hypercorn==0.6.0
quart==0.8.1
rethinkdb